requirements.txt
feedparser
python-frontmatter
python-slugify
requests
tweepy
